feat: 각 page 별로 이벤트 선언 위치 수정
