+++ src/App.jsx
@@ -7,6 +7,11 @@ export const App = ({ targetPage }) => {
   const PageComponent = targetPage ?? NotFoundPage;
   const error = globalStore.getState().error;
 
+  const closeError = (e) => {
+    e.preventDefault();
+    globalStore.setState({ error: null });
+  };
+
   return (
     <div>
       <PageComponent />
@@ -15,6 +20,7 @@ export const App = ({ targetPage }) => {
           id="error-boundary"
           className="fixed bottom-4 left-4 right-4 bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded-lg shadow-lg transition-opacity duration-300 hover:opacity-75"
           role="alert"
+          onClick={closeError}
         >
           <div className="flex justify-between items-center">
             <div>
