+++ src/pages/ProfilePage.jsx
@@ -2,22 +2,49 @@
 import { createVNode } from '../lib';
 import { Footer, Header, Navigation } from '../components/index.js';
 import { globalStore } from '../stores/index.js';
+import { userStorage } from '../storages/index.js';
+import { router } from '../main.jsx';
 
 export const ProfilePage = () => {
   const { loggedIn, currentUser } = globalStore.getState();
   const { username = '', email = '', bio = '' } = currentUser ?? {};
+
+  const updateProfile = (e) => {
+    e.preventDefault();
+    const formData = new FormData(e.target);
+    const updatedProfile = Object.fromEntries(formData);
+    const currentUser = { ...globalStore.getState().currentUser, ...updatedProfile };
+    globalStore.setState({ currentUser });
+    userStorage.set(currentUser);
+    alert('프로필이 업데이트되었습니다.');
+  };
+
+  const logout = (e) => {
+    e.preventDefault();
+
+    globalStore.setState({ currentUser: null, loggedIn: false });
+    router.push('/login');
+    userStorage.reset();
+  };
+
+  const navigate = (e) => {
+    e.preventDefault();
+    if (e.target.tagName !== 'A' || !e.target.dataset.link) return;
+
+    router.push(e.target.href.replace(window.location.origin, ''));
+  };
 
   return (
     <div className="bg-gray-100 min-h-screen flex justify-center">
       <div className="max-w-md w-full">
         <Header />
 
-        <Navigation loggedIn={loggedIn} />
+        <Navigation loggedIn={loggedIn} logout={logout} navigate={navigate} />
 
         <main className="p-4">
           <div className="bg-white p-8 rounded-lg shadow-md">
             <h2 className="text-2xl font-bold text-center text-blue-600 mb-8">내 프로필</h2>
-            <form id="profile-form">
+            <form id="profile-form" onSubmit={updateProfile}>
               <div className="mb-4">
                 <label htmlFor="username" className="block text-gray-700 text-sm font-bold mb-2">
                   사용자 이름
